Moving Files to its right place
