+++ packages/Tabs/SettingTab/Widgets/ClearSourceButton.py
