+++ packages/Tabs/SettingTab/Widgets/ExtensionsCheckableComboBox.py
